// src/app/models/modelAttribute.ts
/**
 * Covid19-case-follow-up
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModelAttribute { 
    attName?: string;
    attType?: string;
    attCode?: string;
    attLabel?: string;
    attDesc?: string;
}