// src/app/models/requestSaveFollowUp.ts
/**
 * Covid19-case-follow-up
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModelEnumIdValue } from './modelEnumIdValue';
import { ModelPuiFollowUpList } from './modelPuiFollowUpList';
import { ModelPuiFollowUp } from './modelPuiFollowUp';
import { RequestSaveCase } from './requestSaveCase';


export interface RequestSaveFollowUp { 
    id?: number;
    caseCode?: string;
    firstName: string;
    lastName: string;
    dob?: string;
    gender?: RequestSaveFollowUp.GenderEnum;
    occupation?: string;
    passportNumber: string;
    incidentContactPhone1: string;
    incidentContactPhone2?: string;
    countryOfResidence: string;
    countryOfOrigin?: string;
    nationality?: string;
    region?: string;
    subcityOrZone?: string;
    woreda?: string;
    kebele?: string;
    houseNo?: string;
    phoneNo?: string;
    latitude?: string;
    presumptiveResult?: ModelEnumIdValue;
    confirmedResult?: ModelEnumIdValue;
    identifiedBy: ModelEnumIdValue;
    parentCaseCode?: string;
    recentTravelTo?: string;
    admittedToFacility?: ModelEnumIdValue;
    status?: ModelEnumIdValue;
    modifiedBy: string;
    reportingDate?: string;
    modifiedDate?: string;
    list: Array<ModelPuiFollowUp>;
}
export namespace RequestSaveFollowUp {
    export type GenderEnum = 'F' | 'M';
    export const GenderEnum = {
        F: 'F' as GenderEnum,
        M: 'M' as GenderEnum
    };
}